refactor: format all files using black formatter
Reformatted the code to improve readability and applied "black" style Python formatting. All code modifications were related to formatting—changing the location of line breaks, adding whitespace around operators, and similar tweaks. This refactoring affects no observable behavior and therefore, no changes in the application's functionality would be noted after the commit. The main goal is to make the code more aesthetically pleasing and easier to maintain in the future.