feat: introduce a new `Coordinates` object
This commit introduces a new dataclass named `Coordinates` which is used throughout the codebase to designate a pair of x and y coordinates. It also updates the import statement of `Generator` to come from the `collections.abc` module for it to be more descriptive in the code.

This `Coordinates` object offers benefits such as better code readability and a more intuitive way of dealing with a pair of x and y coordinates. Furthermore, importing `Generator` from `collections.abc` allows for a more pythonic and consistent way of typing.

Closes #32