refactor: tweak `Coordinates` dataclass
In this commit, the dataclass `Coordinates` has been restructured to use field annotations for `x_coordinate` and `y_coordinate` instead of initializing them in a method. `self._coordinates`, which was previously a tuple holding these coordinates, has been replaced by a property method `_as_tuple` that returns a tuple of the coordinates.

The main reason for these changes is the code readability and optimization. The invocations of the individual coordinates in the `__hash__`, `__iter__` and `__str__` methods are now more explicit, and the need for an extra attribute holding the tuple of coordinates has been eliminated, optimizing memory usage.