refactor: rename variables and methods
This commit refactors variable and method names in the codebase to improve clarity and consistency. It renames variables `swap` to `remaining_swaps` in the `SpellSolver` class to better indicate the purpose of the variable. Additionally, it renames methods `process_node` to `generate_words`, `process_path_aux` to `explore_paths_auxiliary`, `process_path` to `explore_paths`, and `process_gameboard` to `word_list` in the `SpellSolver` class for clearer and more descriptive naming.